# backend/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "tafa3olcard Backend API",
    "version": "1.0.0",
    "description": "API Documentation for the tafa3olcard backend"
  },
  "servers": [
    {
      "url": "http://localhost:5000/api/v1",
      "description": "Development server"
    }
  ],
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "security": [
    {
      "BearerAuth": []
    }
  ],
  "paths": {
...
    "/admin/stocks/products": {
      "get": {
        "summary": "List stock products",
        "tags": [
          "StockProducts"
        ],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Products retrieved"
          }
        }
      },
      "post": {
        "summary": "Create stock product",
        "tags": [
          "StockProducts"
        ],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "responses": {
          "201": {
            "description": "Product created"
          }
        }
      }
    },
    "/admin/stocks/products/import/preview": {
      "post": {
        "summary": "Preview provider products before importing",
        "tags": [
          "StockProducts"
        ],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Provider products retrieved"
          }
        }
      }
    },
    "/admin/stocks/products/import": {
      "post": {
        "summary": "Import provider products into stock products",
        "tags": [
          "StockProducts"
        ],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "responses": {
          "201": {
            "description": "Provider products imported"
          }
        }
      }
    },
    "/admin/stocks/products/{id}/api-connections": {
      "get": {
        "summary": "List product API connections",
        "tags": [
          "StockProducts"
        ],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Product API connections retrieved"
          }
        }
      },
      "post": {
        "summary": "Create product API connection",
        "tags": [
          "StockProducts"
        ],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Product API connection created"
          }
        }
      }
    },
    "/admin/stocks/products/{id}/api-connections/{connectionId}": {
      "patch": {
        "summary": "Update product API connection",
        "tags": [
          "StockProducts"
        ],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "connectionId",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Product API connection updated"
          }
        }
      },
      "delete": {
        "summary": "Soft delete product API connection",
        "tags": [
          "StockProducts"
        ],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "connectionId",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Product API connection deleted"
          }
        }
      }
    },
    "/admin/stocks/products/{id}/api-connections/{connectionId}/activate": {
      "patch": {
        "summary": "Activate product API connection",
        "tags": [
          "StockProducts"
        ],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "connectionId",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Product API connection activated"
          }
        }
      }
    },
    "/admin/stocks/products/{id}": {
      "get": {
        "summary": "Get stock product",
        "tags": [
          "StockProducts"
        ],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Product retrieved"
          }
        }
      },
      "patch": {
        "summary": "Update stock product",
        "tags": [
          "StockProducts"
        ],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Product updated"
          }
        }
      }
    },
...
  },
...
}